Factor out InstructionCard into its own file
- InstructionCard is now in `common/`, not `scroll.dart`.
- InstructionCard simply takes in some instruction text, and displays a
  styled card for that text.
- Remove unnecessary `new` from `explore_menu.dart`.
